Add a changelogs page and other minor fixes (#15)
* Add a changelogs page.

* Fix local search highlight colors.

* Make changelog headers a link.

* Replace GitHub links with local ones.

* Fix missing `Badge` component.

* Show release contributors.

* Add missing flex-wrap to mobile devices.

* Add missing `alt` attribute to the avatar.

* Use `@theme` alias while importing components from the theme.

* Minor design tweaks to the extensions items.

* Show contributors on WhatsNew component as well.

* Fix extensions page not centered.

* Show sidebar on the changelogs page.

* Fix missing link replacements.

* Add missing unmentioned credits.

+++ website/src/.vitepress/config/themeConfig.ts
@@ -59,7 +59,7 @@ const themeConfig: DefaultTheme.Config = {
 	lastUpdated: {
 		text: "Last updated",
 		formatOptions: {
-			dateStyle: "long",
+			dateStyle: "short",
 			timeStyle: "short",
 		},
 	},
